PDF y funciones juntadas adaptadas al spring security

+++ src/main/java/com/tfg/springmarket/model/entities/ProductosEstablecimiento.java
@@ -23,7 +23,7 @@ public class ProductosEstablecimiento implements Serializable {
     private Double precioCoste;
 
     @Column(name = "precio_venta")
-    private BigDecimal precioVenta;
+    private Double precioVenta;
 
     @Column(name = "stock")
     private Integer stock;
